// src/api/generated/schemas/appNotificationType.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * TripBite API
 * OpenAPI spec version: 1.0
 */

export type AppNotificationType = typeof AppNotificationType[keyof typeof AppNotificationType];


export const AppNotificationType = {
  letterreceived: 'letter.received',
  letterliked: 'letter.liked',
  letterdelivered: 'letter.delivered',
  tournamentshared: 'tournament.shared',
  event: 'event',
  security: 'security',
} as const;